サイマルキャストエンコーダーの実装を隠蔽する
